refactor: move preloading & adding icons outside of initilization
